Added code to request irradiance values from sensor
Theres a weird bug though. The c# app takes too long every 60th sample.

+++ Fenrisulfr/Fenrisulfr/SensorResult.cs
@@ -2,8 +2,8 @@ namespace Fenrisulfr
 {
     public struct SensorResult
     {
-        public int Read770;
-        public int Read940;
+        public float Read770;
+        public float Read940;
         public long Milliseconds;
     }
 }
